requirements.txt
blinker==1.9.0
click==8.3.2
et_xmlfile==2.0.0
Flask==3.1.3
gevent==26.4.0
greenlet==3.4.0
gunicorn==25.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
playwright==1.58.0
pyee==13.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
schedule==1.2.2
six==1.17.0
typing_extensions==4.15.0
Werkzeug==3.1.8
zope.event==6.1
zope.interface==8.3